requirements.txt
fastapi==0.118.2
uvicorn[standard]==0.37.0
jinja2==3.1.4
sqlalchemy==2.0.34
psycopg[binary]==3.2.10
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
# Pin bcrypt to avoid incompatibilities with passlib 1.7.4 (bcrypt 4.1+ removed __about__)
bcrypt==4.0.1
python-dotenv==1.0.1
itsdangerous==2.2.0
cloudinary==1.41.0
alembic==1.13.2
slowapi==0.1.9
# Upgrade fastapi-mail and remove pydantic pin for Python 3.13+ compatibility
fastapi-mail==1.3.1
